+++ src/main/java/org/aero/mtip/metamodel/core/AbstractDiagram.java
@@ -12,11 +12,8 @@ import java.awt.Color;
 import java.awt.Point;
 import java.awt.Rectangle;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import javax.annotation.CheckForNull;
@@ -28,6 +25,7 @@ import org.aero.mtip.constants.SysmlConstants;
 import org.aero.mtip.constants.UAFConstants;
 import org.aero.mtip.constants.XmlTagConstants;
 import org.aero.mtip.io.CommonPresentationElement;
+import org.aero.mtip.io.Exporter;
 import org.aero.mtip.io.Importer;
 import org.aero.mtip.metamodel.core.general.Link;
 import org.aero.mtip.profiles.MagicDraw;
@@ -64,12 +62,7 @@ import com.nomagic.uml2.ext.magicdraw.statemachines.mdbehaviorstatemachines.Regi
 
 public abstract class AbstractDiagram extends CommonElement {
   private static Map<String, String> cameoToMtipType = createCameoToMtipMap();
-  private static Collection<String> reversedEndsRelationships = new HashSet<>(Arrays.asList(SysmlConstants.ABSTRACTION,
-      SysmlConstants.CONTROL_FLOW, SysmlConstants.COPY, SysmlConstants.DEPENDENCY, SysmlConstants.EXTEND, SysmlConstants.GENERALIZATION,
-      SysmlConstants.INCLUDE, SysmlConstants.INFORMATION_FLOW, SysmlConstants.INTERFACE_REALIZATION, SysmlConstants.ITEM_FLOW,
-      SysmlConstants.MESSAGE, SysmlConstants.OBJECT_FLOW, SysmlConstants.REFINE, SysmlConstants.SATISFY, SysmlConstants.TRACE,
-      SysmlConstants.TRANSITION, SysmlConstants.USAGE, SysmlConstants.VERIFY));
-
+  
   public HashMap<Element, Rectangle> elementsOnDiagram;
   public List<Element> relationshipsOnDiagram;
   public List<CommonPresentationElement> noElementPresentationElements;
@@ -440,22 +433,31 @@ public abstract class AbstractDiagram extends CommonElement {
     if (pathElement.getSupplierPoint() == null || pathElement.getClientPoint() == null) {
       return;
     }
-
+    
+    CommonRelationship cr = Exporter.getCommonRelationshipsFactory().createElement(relationship);
+    
+    if (cr == null) {
+      Logger.log(String.format("Could not write relationship metadata. No common relationship created from %s", MtipUtils.getEntityType(relationship)));
+      return;
+    }
+    
     org.w3c.dom.Element relDataTag = XmlWriter.createTag(XmlTagConstants.RELATIONSHIP_METADATA, XmlTagConstants.ATTRIBUTE_TYPE_DICT);
 
     org.w3c.dom.Element supplierPointTag = XmlWriter.createTag(XmlTagConstants.SUPPLIER_POINT, XmlTagConstants.ATTRIBUTE_TYPE_DICT);
     org.w3c.dom.Element clientPointTag = XmlWriter.createTag(XmlTagConstants.CLIENT_POINT, XmlTagConstants.ATTRIBUTE_TYPE_DICT);
 
-    writeCoordinates(supplierPointTag, getSupplierPoint(relationship, pathElement));
-    writeCoordinates(clientPointTag, getClientPoint(relationship, pathElement));
+    writeCoordinates(supplierPointTag, cr.getSupplierPoint(pathElement));
+    writeCoordinates(clientPointTag, cr.getClientPoint(pathElement));
 
     org.w3c.dom.Element breakPointsTag = XmlWriter.createTag(XmlTagConstants.BREAK_POINT, XmlTagConstants.ATTRIBUTE_TYPE_LIST);
-
-    for (int i = 0; i < pathElement.getBreakPoints().size(); i++) {
+    
+    List<Point> breakPoints = cr.getBreakPoints(pathElement);
+    
+    for (int i = 0; i < breakPoints.size(); i++) {
       org.w3c.dom.Element breakPointTag = XmlWriter.createTag(XmlTagConstants.BREAK_POINT, XmlTagConstants.ATTRIBUTE_TYPE_DICT);
       XmlWriter.addAttributeKey(breakPointTag, Integer.toString(i));
 
-      writeCoordinates(breakPointTag, pathElement.getBreakPoints().get(i));
+      writeCoordinates(breakPointTag, breakPoints.get(i));
 
       XmlWriter.add(breakPointsTag, breakPointTag);
     }
@@ -532,22 +534,6 @@ public abstract class AbstractDiagram extends CommonElement {
     }
 
     return MagicDraw.hasCustomImageHolderStereotype(presentationElement.getElement());
-  }
-
-  protected static Point getSupplierPoint(Element relationship, PathElement pathElement) {
-    if (AbstractDiagram.reversedEndsRelationships.contains(MtipUtils.getEntityType(relationship))) {
-      return pathElement.getClientPoint();
-    }
-
-    return pathElement.getSupplierPoint();
-  }
-
-  protected static Point getClientPoint(Element relationship, PathElement pathElement) {
-    if (AbstractDiagram.reversedEndsRelationships.contains(MtipUtils.getEntityType(relationship))) {
-      return pathElement.getSupplierPoint();
-    }
-
-    return pathElement.getClientPoint();
   }
 
   private static Map<String, String> createCameoToMtipMap() {