+++ src/main/java/org/aero/mtip/metamodel/core/general/Generalization.java
@@ -17,5 +17,6 @@ public class Generalization extends CommonDirectedRelationship {
 		this.metamodelConstant = SysmlConstants.GENERALIZATION;
 		this.xmlConstant = XmlTagConstants.GENERALIZATION;
 		this.element = f.createGeneralizationInstance();
+		this.isReversedEndsRelationship = true;
 	}
 }