+++ src/main/java/org/aero/mtip/metamodel/sysml/requirements/Satisfy.java
@@ -1,40 +1,24 @@
-/* The Aerospace Corporation MTIP_Cameo
-Copyright 2022 The Aerospace Corporation
-
-This product includes software developed at
-The Aerospace Corporation (http://www.aerospace.org/). */
+/*
+ * The Aerospace Corporation MTIP_Cameo Copyright 2022 The Aerospace Corporation
+ * 
+ * This product includes software developed at The Aerospace Corporation
+ * (http://www.aerospace.org/).
+ */
 
 package org.aero.mtip.metamodel.sysml.requirements;
 
 import org.aero.mtip.constants.SysmlConstants;
 import org.aero.mtip.constants.XmlTagConstants;
-import org.aero.mtip.metamodel.core.CommonDirectedRelationship;
-import org.aero.mtip.util.ElementData;
+import org.aero.mtip.metamodel.core.general.Abstraction;
+import org.aero.mtip.profiles.SysML;
 
-import com.nomagic.magicdraw.core.Project;
-import com.nomagic.uml2.ext.jmi.helpers.StereotypesHelper;
-import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Element;
-import com.nomagic.uml2.ext.magicdraw.mdprofiles.Profile;
-import com.nomagic.uml2.ext.magicdraw.mdprofiles.Stereotype;
+public class Satisfy extends Abstraction {
 
-public class Satisfy extends CommonDirectedRelationship {
-
-	public Satisfy(String name, String importId) {
-		super(name, importId);
-		this.creationType = XmlTagConstants.ELEMENTS_FACTORY;
-		this.metamodelConstant = SysmlConstants.SATISFY;
-		this.xmlConstant = XmlTagConstants.SATISFY;
-		this.element = f.createAbstractionInstance();
-	}
-
-	@Override
-	public Element createElement(Project project, Element owner, Element client, Element supplier, ElementData xmlElement) {
-		super.createElement(project, owner, client, supplier, xmlElement);
-		
-		Profile sysml = StereotypesHelper.getProfile(project, "SysML");
-		Stereotype satisfyStereotype = StereotypesHelper.getStereotype(project,  "Satisfy", sysml);
-		StereotypesHelper.addStereotype(element, satisfyStereotype);
-
-		return element;
-	}
+  public Satisfy(String name, String importId) {
+    super(name, importId);
+    this.creationType = XmlTagConstants.ELEMENTS_FACTORY;
+    this.metamodelConstant = SysmlConstants.SATISFY;
+    this.xmlConstant = XmlTagConstants.SATISFY;
+    this.creationStereotype = SysML.getSatisfyStereotype();
+  }
 }